Re-write RRDP collector for archives.
