+++ src/collector/mod.rs
@@ -31,3 +31,4 @@ mod base;
 mod rrdp;
 mod rsync;
 
+pub mod newrrdp;
