+++ src/collector/rrdp/mod.rs
@@ -1,0 +1,11 @@
+#![allow(dead_code)]
+
+pub use self::base::{Collector, LoadResult, ReadRepository, Run};
+pub use self::http::HttpStatus;
+pub use self::update::SnapshotReason;
+
+mod archive;
+mod base;
+mod http;
+mod update;
+
