+++ src/utils/sync.rs
@@ -2,7 +2,6 @@
 
 use std::sync::{Mutex as StdMutex, RwLock as StdRwLock};
 
-
 pub use std::sync::{MutexGuard, RwLockReadGuard, RwLockWriteGuard};
